fix: display passed alarms for futures events
Admit my vdirsyncer instance just fetched an ics event from a futur
event in 1 hour, and a passed alarm from 1 hour.
Due to the fact this alarm has not been done, we want it to be displayed.

This fix got performance optimization consequences. Maybe we could
daemonize remhind on minutes instead of 45s?